Cinnamon - Add a pdf viewer by default
Atril is the pdf viewer of the mate desktop, the Cinnamon spin already include the
Eye of Mate app so include this app will not requiere to much more dependencies to
be included.

Not having a pdf viewer by default it is bad experience for user in Live mode or
doing clean install of the Cinnamon Spin.

References:
 - https://github.com/mate-desktop/atril

Signed-off-by: William Moreno Reyes <williamjmorenor@fedoraproject.org>

 En la rama cinnamon-pdf
 Cambios a ser confirmados:
	modificado:     fedora-cinnamon-common.ks