Cinnamon - Add a music player
Yarock is a music player not specific to any desktop enviroment, since Fedora now
have support to more multimedia formats it is a nice to new user to have a nice
looking music player included in the default Cinnamon Spin.

Yarock is based in QT, the iso image for Cinnamon in F32 is around 1.4 GB, so
there is space to include more apps by default.

References:
 - https://seb-apps.github.io/yarock/

Signed-off-by: William Moreno Reyes <williamjmorenor@fedoraproject.org>