Cleanup fedora-live-minimization some more
mpage is only 60K in size and requires only libc and rtld. Seems
worthless to drop it.

isdn4k-utils doesn't exist any more.

For the sane stuff: nothing sane-related is pulled into most
images any more. cinnamon-desktop, design-suite and
gnome-desktop list some SANE packages, but that clearly means
they *want* to include them, and the kickstarts for those
spins don't include fedora-live-minimization anyway. The
'graphics' group lists xsane-gimp conditional on gimp being
included, but no desktop environment or kickstart includes the
'graphics' group as a default, so it won't be on any live images.

This leaves only hplip in the file. That *is* of significant
size and *would* be pulled into most of the spins which include
this file, so we can't drop it yet at least. There may be an
argument for dropping it from comps with the current state of
IPP Everywhere support, I guess.

Signed-off-by: Adam Williamson <awilliam@redhat.com>