Disable plasma-welcome on KDE lives for F37
This is the same as https://pagure.io/livesys-scripts/c/bd45f85
but in kickstarts, because we don't use livesys-scripts in F37.
This is needed because the newer Plasma was sent to F37 as well,
so the F37 respin KDE live shows the welcome tour now:

https://openqa.fedoraproject.org/tests/1892872#step/desktop_notifications/16

it shouldn't do that, the tour should only be shown post-install.

Signed-off-by: Adam Williamson <awilliam@redhat.com>