Revert "Turn on anaconda-webui for workstation"
This reverts commit ce5b31416f0bef6ff70acc3363088a95e2ee5983.
The anaconda update with the webUI changes is failing tests,
and we do not have time to resolve this right away. We can't
leave this in place without also pushing the anaconda update
stable and adjusting the openQA tests, so let's revert it for
now so tests pass on other updates, until we can come back and
clean up the webUI stuff tomorrow.