container-toolbox: Use default-editor, not nano-default-editor
In Fedora 33, nano-default-editor was introduced to make GNU nano the
default editor [1].  After a year, it was found out that it breaks
upgrades across Fedora releases if the user had chosen some other editor
as the default [2].

The difference doesn't matter here because the fedora-toolbox OCI image
is being created from scratch without involving any upgrades.  However,
since fedora-comps was switched to default-editor [3], it's better to
do the same and stay consistent.  If nothing else, it won't make the
reader stop and ponder if there's any real reason that it was done
differently for the fedora-toolbox image.

[1] fedora nano commit cf1ce689f154d0a1
    https://src.fedoraproject.org/rpms/nano/c/cf1ce689f154d0a1
    https://src.fedoraproject.org/rpms/nano/pull-request/1
    https://fedoraproject.org/wiki/Changes/UseNanoByDefault

[2] fedora nano commit e53ee57baa33a1d8
    https://src.fedoraproject.org/rpms/nano/c/e53ee57baa33a1d8
    https://src.fedoraproject.org/rpms/nano/pull-request/7
    https://bugzilla.redhat.com/show_bug.cgi?id=1955884

[3] fedora-comps commit 3496ed634cf452da
    https://pagure.io/fedora-comps/c/3496ed634cf452da
    https://pagure.io/fedora-comps/pull-request/663
    https://bugzilla.redhat.com/show_bug.cgi?id=1955884

https://pagure.io/fedora-kickstarts/pull-request/992