Only base containers should have tzdata
The base container should always install tzdata to ensure that it is
available for applications built on top of the base container.

The minimal container should never have tzdata installed, and the
application should install it as part of the application dependencies.

Starting with Fedora 39 we have the capability to remove tzdata from
the minimal images without resorting to deleting files:
https://fedoraproject.org/wiki/Changes/AllowRemovalOfTzdata