container-toolbox: Restore the tests
The tests can't be run inside the '%post --nochroot' section because
the paths being tested don't include the /mnt/sysimage prefix.  The
tests must be run in the %post section without --nochroot so that they
are run inside the chroot environment.

The translations for the RPM manuals were removed upstream during the
RPM 4.19 development cycle [1].  So, replace them with rpm2cpio(8),
which is another popular command shipped by the rpm package.

[1] RPM commit 4df74a9644b18136
    https://github.com/rpm-software-management/rpm/commit/4df74a9644b18136
    https://github.com/rpm-software-management/rpm/pull/2245

https://pagure.io/fedora-kickstarts/pull-request/1008