fedora-container-toolbox: drop systemd
Since change 48e2c3b5598de32f6dabfb5d1985d17aeab97d58 this kickstart
is pulling in systemd.

This was noticed because since
b5fc5fd17d204870eebaa3f48f73e9f3c6d00697 started bringing in
weak-dependencies, we started installing systemd-resolved is which
created a symlinked /etc/resolv.conf in the image.  Toolbox will not
currently reset this on container start, as it is a symlink (this
behaviour is a bit complicated; see [1]).  This leads to an
incompatability running the toolbox on *non* systemd-resolved hosts
(e.g. RHEL9); you are left with a dangling symlink and no
name-resolution in the toolbox.

We do not want systemd in the toolbox image by default it; remove it
from the list.  Exclude systemd-resolved specifically, so if something
else brings in systemd we still don't include this.

[1] https://github.com/containers/toolbox/issues/1410

https://pagure.io/fedora-kickstarts/pull-request/1027