container-toolbox: Add a fallback test to catch missing RPM content
This was added to the Container/Dockerfile equivalent of the
fedora-toolbox OCI images in recent times [1].

Currently, there's a hand-maintained list of files that are used as a
smoke test to ensure that RPM content, like documentation and
translations, didn't get stripped out of the fedora-toolbox image.  This
list is merely a smoke test and can never cover every single file.  So,
it's better to have something more dynamic that can catch any content
that went missing from any RPM in the image.

It should be possible to drop the test with the hand-maintained list,
once it's clear that this new test is really working as intended.

[1] Toolbx commit 154156d86036624f
    https://github.com/containers/toolbox/commit/154156d86036624f
    https://github.com/containers/toolbox/issues/603

https://pagure.io/fedora-kickstarts/pull-request/1028