container-toolbox: Don't try to copy a non-existent file
There's no README.md to copy, and it polluted the build logs with a
needless error message:
  INFO anaconda:program: cp: cannot stat '/README.md': No such file or
      directory

The Container/Dockerfile equivalent of the fedora-toolbox OCI images
have a /README.md because it used to be required by the Fedora Container
Guidelines [1].  However, that's no longer the case [2] and neither the
fedora nor the fedora-minimal base images have it.  Therefore, it's not
worth bothering with it.

[1] https://docs.fedoraproject.org/en-US/containers/

[2] Fedora Container Docs commit 4fc6827a23ddc265
    https://github.com/containers/docs/commit/4fc6827a23ddc265
    https://github.com/containers/docs/pull/5

Fallout from 48e2c3b5598de32f6dabfb5d1985d17aeab97d58

https://pagure.io/fedora-kickstarts/pull-request/1028