container-toolbox: Exclude glibc-langpack-en
The fedora-toolbox OCI image already includes the glibc-all-langpacks
RPM that ships a pre-built /usr/lib/locale/locale-archive file with all
the locale definitions built in.

The glibc-langpack-en RPM provides definitions for all the en_* locales
in per-locale sub-directories in /usr/lib/locale.  This is a subset of
glibc-langpack-en, even if it uses a different representation, and hence
not necessary at all.

For some reason, glibc-langpack-en is present in the fedora-toolbox:39
image, but not in the images for Fedora 40 and 41.  Therefore, it's
better to be explicit about it as a safeguard against any inadvertent
changes.

https://pagure.io/fedora-kickstarts/pull-request/1028