maintainers: Reassign Games lab to myself and q5sys
The previous maintainers have been inactive on the Games SIG and
lab spin for a couple of years and the most recent keepalive was
answered by JT Pennington.

Thus, reassign to myself as de facto Games SIG chair and JT Pennington
as the last responder for the lab spin.

Reference: https://pagure.io/fedora-pgm/schedule/issue/160